fix: 重命名插件页面
